refactor to strategy

+++ src/main/java/strategy/CalculadorLlamada.java
@@ -1,5 +1,5 @@
 package strategy;
 
 public interface CalculadorLlamada {
-    float costo(float duracion);
+    double costo(double duracion);
 }
